refactor(GOD_OBJECT): extract legacy strategy machine to routing_strategies
ModelRouter mixed two responsibilities: provider registration / metrics /
circuit-breaking, and a six-strategy selection machine (round-robin,
least-busy, task/cost/performance-optimized, load-balanced) that only runs
as the no-tier / TierRouter-failure fallback. Move RoutingStrategy,
RoutingContext, ModelMetrics and the selection methods into a new
routing_strategies.py behind a RoutingStrategySelector; ModelRouter now owns
a selector and delegates _apply_routing_strategy to it. The sole external
importer (model_command.py) is updated to the new module.
